+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v2.1.0
+
+Features:
+
+- Sharec skip fields removed by user (#78)
+
+Fix:
+
+- Aliases for configs pipes (#107)
+
 # v2.0.3
 
 Fix: